Final presentación

+++ app/Http/Controllers/SeguimientoController.php
@@ -11,6 +11,7 @@ use App\Models\Process;
 use App\Models\processCriterio;
 use App\Models\Rol;
 use App\Models\Seguimiento;
+use App\Models\seguimientoPropuesto;
 use Illuminate\Http\Request;
 use Flash;
 use Response;
@@ -27,8 +28,25 @@ class SeguimientoController extends AppBaseController
      */
     public function index($id, $id2, Request $request)
     {
+        $procesosPriorizados = matrizPriorizado::where('process_map_id','=',$id2)
+        ->first();
+        if (empty($procesosPriorizados)) {
+            Flash::error("Completar la priorización de procesos críticos");
+            return redirect(route('processCriterios.index', [$id, $id2]))
+            ->with('company_id',$id)->with('process_map_id',$id2);
+        }
+        $ids = json_decode($procesosPriorizados->process_id_data_seguimiento);
+        $idsAll = json_decode($procesosPriorizados->process_id_data_all);
         if($request->ajax()){
-            $seguimientos = Process::where('process_map_id','=',$id2)->has('seguimientos')->get();
+
+            $seguimientos = Process::whereIn('id', $idsAll)
+            //->where('matriz_priorizado_id','=',$procesosPriorizados->id)
+            ->has('seguimientos')
+            ->whereHas('seguimientos', function($q) use ($id2){
+                $procesosPriorizados = matrizPriorizado::where('process_map_id','=',$id2)->first();
+                $q->where('matriz_priorizado_id', '=',$procesosPriorizados->id);
+            })
+            ->get();
             return DataTables::of($seguimientos)
             ->addColumn('company_id',function($seguimiento){
                 return $seguimiento->processMap->company_id;
@@ -41,10 +59,9 @@ class SeguimientoController extends AppBaseController
             ->rawColumns(['action','actionPropuesto'])
             ->make(true);
         }
-        $procesosPriorizados = matrizPriorizado::where('process_map_id','=',$id2)->first();
-        $ids = json_decode($procesosPriorizados->process_id_data);
-        //$procesosSinHojaCaracterizacion = Process::find($ids);
-        $procesosSinSeguimiento = Process::whereIn('id', $ids)->doesnthave('seguimientos')->whereNull('parent_process_id')->get();
+
+        $procesosSinSeguimiento = Process::whereIn('id', $idsAll)
+        ->whereNull('parent_process_id')->get();
         /*$procesosSinSeguimiento = Process::where('process_map_id','=',$id2)->doesnthave('seguimientos')->get();*/
         $criterios = Criterio::where('process_map_id','=',$id2)->get()->count();
         $procesos = Process::where('process_map_id','=',$id2)->whereNull('parent_process_id')->get()->count();
@@ -63,9 +80,19 @@ class SeguimientoController extends AppBaseController
     }
     public function show($id, $id2, $id3, Request $request)
     {
+        $process = Process::find($id3);
+        $procesosPriorizados = matrizPriorizado::where('process_map_id','=',$id2)
+        ->first();
+        if (empty($procesosPriorizados)) {
+            Flash::error("Completar la priorización de procesos críticos");
+            return redirect(route('processCriterios.index', [$id, $id2]))
+            ->with('company_id',$id)->with('process_map_id',$id2);
+        }
         if($request->ajax()){
             /** @var Seguimiento $seguimientos */
-            $seguimientos = Seguimiento::where('process_id','=',$id3)->get();
+            $seguimientos = Seguimiento::where('process_id','=',$id3)
+            ->where('matriz_priorizado_id','=',$procesosPriorizados->id)
+            ->get();
             return DataTables::of($seguimientos)
             ->addColumn('company_id',function($seguimiento){
                 return $seguimiento->process->processMap->company_id;
@@ -104,22 +131,34 @@ class SeguimientoController extends AppBaseController
         }
         $roles = Rol::where('process_map_id','=',$id2)->get();
         //return $procesosSinSeguimiento;
-        return view('seguimientos.show',compact('roles'))->with('company_id',$id)
+        return view('seguimientos.show',compact('roles','process'))->with('company_id',$id)
         ->with('process_map_id',$id2)
         ->with('process_id',$id3);
     }
     public function getSeguimiento($id, $id2, $id3)
     {
-        $actividades = Seguimiento::where('process_id','=',$id3)->get();
+        $process = Process::find($id);
+        $procesosPriorizados = matrizPriorizado::where('process_map_id','=',$process->process_map_id)->first();
+        $actividades = Seguimiento::where('process_id','=',$id3)->where('matriz_priorizado_id',$procesosPriorizados->id)->get();
         //return $actividades;
         return view('seguimientos.view_seguimiento.seguimiento_table',compact('actividades'))->with('company_id',$id)->with('process_map_id',$id2)->with('seguimiento_id',$id3);
     }
 
-    public function storeActivity(CreateSeguimientoRequest $request)
+    public function storeActivity(Request $request)
     {
         try{
             $input = $request->all();
+            $procesosPriorizados = matrizPriorizado::where('process_map_id','=',$request->process_map_id)
+            ->first();
+            $input['matriz_priorizado_id'] = $procesosPriorizados->id;
             $seguimiento = Seguimiento::create($input);
+            $ids = json_decode($procesosPriorizados->process_id_data_seguimiento); //Ids actual
+            //Remover proceso de la lista actual
+            if (($key = array_search($request->process_id, $ids)) !== false) {
+                unset($ids[$key]);
+            }
+            $procesosPriorizados->process_id_data_seguimiento = json_encode(array_values($ids));
+            $procesosPriorizados->save();
             return Response::json(['status'=>'200'], 200);
         }
         catch(Exception $e){
@@ -128,14 +167,17 @@ class SeguimientoController extends AppBaseController
     }
     public function getTimes($id)
     {
+        $process = Process::find($id);
+
+        $procesosPriorizados = matrizPriorizado::where('process_map_id','=',$process->process_map_id)->first();
         try{
-            $rolTimes = Seguimiento::where('process_id','=',$id)
+            $rolTimes = Seguimiento::where('process_id','=',$id)->where('matriz_priorizado_id',$procesosPriorizados->id)
             ->selectRaw('rol.name as name, sum(time) as total')
             ->join('process','process.id','=','seguimiento.process_id')
             ->join('rol','rol.id','=','seguimiento.rol_id')
             ->groupBy('rol.name')
             ->get();
-            $flowTimes = Seguimiento::where('process_id','=',$id)
+            $flowTimes = Seguimiento::where('process_id','=',$id)->where('matriz_priorizado_id',$procesosPriorizados->id)
             ->selectRaw('seguimiento.flow_id, sum(time) as total')
             ->join('process','process.id','=','seguimiento.process_id')
             ->join('rol','rol.id','=','seguimiento.rol_id')
@@ -207,8 +249,27 @@ class SeguimientoController extends AppBaseController
         //Eliminar todos los seguimientos del proceso
         //return $id3;
         $proceso = Process::where('id','=',$id3)->get();
+
+
+
+
+        $procesosPriorizados = matrizPriorizado::where('process_map_id','=',$id2)->first();
+        if (empty($procesosPriorizados)) {
+            Flash::error("Completar la priorización de procesos críticos");
+            return redirect(route('processCriterios.index', [$id, $id2]))->with('company_id',$id)->with('process_map_id',$id2);
+        }
+        $ids = json_decode($procesosPriorizados->process_id_data_seguimiento); //Ids actual
+        array_unshift($ids,strval($id3));
+        //return json_encode(array_values($ids));
+        $procesosPriorizados->process_id_data_seguimiento = json_encode(array_values($ids));
+        $procesosPriorizados->save();
+
+
+
+
         //return $proceso;
         $seguimiento = Seguimiento::where('process_id','=',$id3)->delete();
+        $seguimientoPropuesto = seguimientoPropuesto::where('process_id','=',$id3)->delete();
 
         if (empty($seguimiento)) {
             Flash::error(__('messages.not_found', ['model' => __('models/seguimientos.singular')]));
